PostgreSQL_Citrus/CQRS/requirements.txt
psycopg2-binary==2.9.9
pika==1.3.2